Add priority reliable queue
So that a single worker can process work from multiple queues of
different priority (e.g. a "critical" queue, a default queue and a
"backlog" queue), add a new Priority Reliable Queue class which will
check a given array of Redis keys for work in a specific order.

This will ensure that the highest priority queue is empty before
checking the next queue for work and so on until the last queue.