Weight priority queues
Instead of waiting for each queue to drain in priority order, rely on
probability to pull from ordered queues in rough priority.

This way, we can block for a second when waiting for work to alleviate
pressure on Redis when queues are empty and we can avoid queue
starvation by checking all queues even if the highest priority is always
full.